src/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-multipart==0.0.22
python-dotenv==1.2.2
pydantic-settings==2.13.1
aiofiles==25.1.0
SQLAlchemy==2.0.48
asyncpg==0.31.0
alembic==1.18.4
google-cloud-storage==3.9.0
redis==7.3.0
cryptography==46.0.6
bcrypt==5.0.0
python-jose==3.5.0
passlib==1.7.4
sqlalchemy-upsert-kit==0.1.1
email-validator==2.3.0
httpx==0.28.1
websockets==16.0
